Rename iosApp to Linkding.
