+++ app/ios/Linkding/Assets.xcassets/AccentColor.colorset/Contents.json
