+++ app/ios/Linkding/Assets.xcassets/AppIcon.appiconset/AppIcon.png
