+++ app/ios/Linkding/Assets.xcassets/AppIcon.appiconset/Contents.json
