+++ app/ios/Linkding/Assets.xcassets/Contents.json
