+++ app/ios/Linkding/Info.plist
