--- app/ios/iosApp/iOSApp.swift
@@ -1,37 +1,0 @@
-import SwiftUI
-import LinkdingKt
-
-class AppDelegate: UIResponder, UIApplicationDelegate {
-
-    lazy var applicationComponent: IosAppComponent = createApplicationComponent(
-        appDelegate: self
-    )
-
-}
-
-@main
-struct iOSApp: App {
-    @UIApplicationDelegateAdaptor(AppDelegate.self) var delegate
-	var body: some Scene {
-		WindowGroup {
-            let uiComponent = createUiComponent(
-                appComponent: delegate.applicationComponent
-            )
-			ContentView(component: uiComponent)
-		}
-	}
-}
-
-private func createApplicationComponent(
-    appDelegate: AppDelegate
-) -> IosAppComponent {
-    return IosAppComponent.companion.create()
-}
-
-private func createUiComponent(
-    appComponent: IosAppComponent
-) -> IosUiViewControllerComponent {
-    return IosUiViewControllerComponent.companion.create(
-        appComponent: appComponent
-    )
-}
